backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-socketio==5.3.6
anthropic>=0.40.0
python-dotenv==1.0.1
gunicorn==22.0.0
reportlab==4.2.2
markdown==3.6
eventlet==0.36.1
